[nvim] Convert vim configuration into neovim
This commit changes the default use case from being vim to neovim. This
is a reflection of my current use of vim. I am using neovim as my
default editor and vim is a fallback.

There is a new structure to how the configuration is setup. Plugins are
defined in the `bundles` folder. A plugin has the Plug definition and
configuration all in one place.

All core vim settings are defined in core.

Autoloaded functions are used to define operations and util functions