feat(brn): Add bulk rename script
Wrote a script to take either a list of files or the current directory
and open a buffer in your editor  (vim) from there you can bunk rename
the files. Once file is saved will loop though and rename files.

If file is tracked with git it will move with git, else with default
system mv