[nvim] Install packages with packer when installed
Nvim tries to add packer. If it is not installed it will git clone
packer into:

    $XDG_CACHE_DIR/nvim/site/pack/packer/opt/packer.nvim

It will then readd this path into the runtimepath with packadd and setup
a autocmd to call `:PackerInstall`.

When adding nvim configuration to a clean machine it will auto install
packer packages.

Note: There seams to be an issue with packer failes to install packer
because it already exists. Not sure how to handle this currently. Also
for some reason fzf install does not work as well but when nvim is
restarted it works fine.