[nvim] Move coc.vim from after to plugin
This change was done as I thought that moving it from lua to an
`after/plugin` file would make coc install to '~/.cache/nvim'. This was
true for linux but not for windows. The installation dir was in Appdata.
I want to mirror linux as much as possible so this change was made.